+++ src-docs/App.vue
@@ -371,6 +371,7 @@ import issue65 from './json/issues/issue-65.json'
 import issue33 from './json/issues/issue-33.json'
 import issue31 from './json/issues/issue-31.json'
 import issue78 from './json/issues/issue-78.json'
+import issue79 from './json/issues/issue-79.json'
 import joditAsProp from './json/issues/jodit-as-prop.json'
 import perfCheckbox from './json/issues/perf-checkbox.json'
 import perfOneOf from './json/issues/perf-oneOf.json'
@@ -573,6 +574,7 @@ export default {
           'issue/issue-64': issue64,
           'issue/issue-65': issue65,
           'issue/issue-78': issue78,
+          'issue/issue-79': issue79,
           'issue/jodit-as-prop': joditAsProp,
           'issue/perf-checkbox': perfCheckbox,
           'issue/perf-oneOf': perfOneOf,